EduSetu/requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.8.2
requests==2.31.0
python-dotenv==1.0.0
cachetools==5.3.1
pyyaml==6.0.1